update tests, need to document the test clases
