+++ test/cases_matching_callbacks/2.btx_callbacks.yaml
@@ -1,0 +1,9 @@
+- :name: usr
+  :regex: \S*_1
+  :args:
+  - :name: _timestamp
+    :type: int64_t
+  - :name: cf_1
+    :type: const char *
+  - :name: pf_1
+    :type: uint64_t
