+++ test/cases_matching_callbacks/2.btx_log.txt
@@ -1,0 +1,4 @@
+[23:20:02.668833240] (+?.?????????) event_1: { cf_1 = "test", pf_1 = 0 }
+[23:20:02.668837077] (+0.000003837) event_1: { cf_1 = "test", pf_1 = 1 }
+[23:20:02.669013249] (+0.000176172) event_1: { cf_1 = "test", pf_1 = 2 }
+[23:20:02.669019918] (+0.000006669) event_1: { cf_1 = "test", pf_1 = 3 }
