+++ test/cases_matching_callbacks/2.btx_model.yaml
@@ -1,0 +1,18 @@
+:stream_classes:
+- :name: sc
+  :default_clock_class: {}
+  :event_common_context_field_class:
+    :type: structure
+    :members:
+      - :name: cf_1
+        :field_class:
+          :type: string
+  :event_classes:
+  - :name: event_1
+    :payload_field_class:
+      :type: structure
+      :members:
+      - :name: pf_1
+        :field_class:
+          :type: integer_unsigned
+          :field_value_range: 64
